add new file for weights and import
